Crear HomeController para a implementar a método addCart

+++ src/main/java/com/maxi/backstore/entities/Product.java
@@ -36,7 +36,7 @@ public class Product {
     private Long id;
     private String name;
     private String description;
-    private Double stock;
+    private Integer stock;
     private Double price;
 
     @Enumerated(EnumType.STRING)
